requirements.txt
Django>=5.2.4
django-allauth>=0.60.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
django-filter>=23.0
Pillow>=10.0.0
requests>=2.31.0
PyJWT>=2.8.0
cryptography>=41.0.0
python-decouple>=3.8
gunicorn>=21.0.0
whitenoise>=6.5.0
psycopg2-binary>=2.9.0
redis>=4.5.0
celery>=5.3.0
django-cors-headers>=4.0.0
djangorestframework>=3.14.0
